requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
oracledb==1.4.1
Jinja2==3.1.2
python-gitlab==3.15.0
pyyaml==6.0.1
marshmallow==3.20.1
gunicorn==21.2.0
structlog==23.1.0
pytest==8.2.0
pytest-cov==5.0.0
cachetools==5.3.3
flasgger==0.9.7.1
